server/requirements.txt
pynput
pyzmq
